Extracted resource definition to generic module
